+++ CHANGELOG.md
@@ -4,6 +4,15 @@ All notable changes will be documented in this file.
 
 Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) principles.
 
+## [2.0.0] - 2022-04-13
+
+### Added
+- Added woff and woff2 formats
+
+### Changed
+- Updated font face params
+- Renamed theme.fontPaths to theme.fontSrc
+
 ## [1.0.1] - 2022-04-06
 
 ### Changed